Changing two types in transaction2
Switched ItemNum to int because if we go with serial, we're going to get into higher numbers than we need (see issue.) Also changing ProductNum default to 1; if it shows up in the transactions index, it's probably going to be one. This should also allow us to allow for negative ProductNum values.